Gestión de estados y ABM

+++ InventarioyStock/src/main/java/com/IntegradorIO/InventarioyStock/OrdenCompra/OrdenCompraRepository.java
@@ -2,5 +2,6 @@ package com.IntegradorIO.InventarioyStock.OrdenCompra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrdenCompraRepository extends JpaRepository<OrdenCompra, Long> {
+public interface OrdenCompraRepository extends JpaRepository<OrdenCompra, Integer> {
+
 }
